Multiple-create: Request Validation (#2057)
## Why make this change?
With support for multiple-create, we introduce additional flexibility in
the GraphQL schema where we make the referencing fields in an entity
optional because their value can be derived from insertions in the
referenced entity. So to say, _the values of referencing fields in a
referencing entity can have 3 sources of truth_. The value for a
referencing field can be derived via:
1. User input data
2. Insertion in the parent referenced entity (if current entity is a
referencing entity for a relationship with its parent entity)
3. Insertion in the child referenced entity (if the current entity is a
referencing entity for a relationship with its child entity)

**_They can all exist at the same time, or none of them can. Presence of
more than one sources of truth leads to possible conflicting values for
the referencing column. Thus, at a time, we want exactly one of them to
be present - so that we exactly one source of truth_**. This needs us to
do request validation during request execution, so that only valid GQL
mutations reach the query generation/query execution stage. This PR adds
the logic for the same.


## What is this change?

- Added class `MultipleMutationInputValidator.cs` to perform the
validations for a multiple-create mutation.
- Added class `MultipleMutationEntityInputValidationContext` to store
relevant information while performing validations on a particular
entity.
- A call to the method
`MultipleMutationInputValidator.ValidateGraphQLValueNode()` will be made
_just after Authorization_ - request validation only occurs when the
user is authorized to execute the mutation.
- The method `MultipleMutationInputValidator.ValidateGraphQLValueNode()`
makes recursive call to itself to parse out the entire create mutation
request body based on an entity or relationship basis. Meaning, this
method will be called on the top-level entity, followed by all the
relationships for that entity, followed by all the relationships for the
entity's relationships and so on....
- The actual validation for the input data for an entity (top-level
entity or a nested entity) which includes column fields and relationship
fields is done by the method
`GraphQLRequestValidator.ValidateObjectFieldNodes()` which is called
from the method `ValidateGraphQLValueNode()`.
- The method `ValidateObjectFieldNodes()` contains logic to ensure that
when the current entity acts as the referencing entity for the
relationships with its target entity, then the current entity should not
specify a value for the referencing fields - as the values for the
referencing fields will be derived from the insertion in the referenced
target entity. Note: The determination of referencing entity is done via
a call to `MultipleCreateOrderHelper.GetReferencingEntityName()` method
added via: https://github.com/Azure/data-api-builder/pull/2056.
- The method `ValidateObjectFieldNodes()` makes calls to
1. `ValidateAbsenceOfReferencingColumnsInTargetEntity()`: To validate
that the current entity does not specify value for referencing columns
to its parent entity (when parent entity is the referenced entity).
2. `ProcessRelationshipField()`: To process a relationship field and
does several things. Please refer to method summary.
3. `ValidateAbsenceOfRepeatedReferencingColumn()`: This method is called
for all the relationships with the current entity which are included in
the request body. This ensures that referencing columns in the
referencing entity (whichever entity, either source or target turns out
as the referencing entity) does not hold references to multiple columns
in the referenced entity - to avoid multiple sources of truth for the
referencing column's value. (Issue:
https://github.com/Azure/data-api-builder/issues/2019)
4. `ValidatePresenceOfRequiredColumnsForInsertion()`: To ensure that we
have one source of truth for values for referencing fields.
5. `ValidateRelationshipFields()`: Recursively perform the same
validations for all the relationships with the current entity included
in the mutation request body.
`
## How was this tested?
- [x] Integration Tests - Added tests to
`MultipleMutationIntegrationTests.cs` to validate:
1. Throwing exception for absence of source of truth for referencing
column for create one mutations
2. Throwing exception for absence of source of truth for referencing
column for create multiple mutations.
3. Throwing exception for presence of multiple sources of truth for
referencing column for create one mutations.
4. Throwing exception for presence of multiple sources of truth for
referencing column for create one mutations.

## Sample Request(s)
1. Request:

![image](https://github.com/Azure/data-api-builder/assets/34566234/b6ef03bf-016d-4a45-9adf-dbe422683f73)
    Response:
    

![image](https://github.com/Azure/data-api-builder/assets/34566234/cbec3e22-3a9e-4bfd-abc8-0fb4fd45cfdd)

2. Request:

![image](https://github.com/Azure/data-api-builder/assets/34566234/6c58f6b9-54b8-42b3-8945-39272023c2bf)

Response:

![image](https://github.com/Azure/data-api-builder/assets/34566234/b0d5f689-d029-4ba9-9aeb-40fca7eb1b85)

3. Request and response:

![image](https://github.com/Azure/data-api-builder/assets/34566234/c39ad294-412c-45d9-9fdd-5efb69908836)

---------

Co-authored-by: Shyam Sundar J <shyamsundarj@microsoft.com>
Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>