Validate relationship field names in config as backing columns (#2175)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/2166
Related to https://github.com/Azure/data-api-builder/pull/2139

## What is this change?

When we validate the relationship fields in the config, ie: source
fields and target fields, we use the backing column names to validate
against. Previously we were using aliases if one existed, but instead we
will validate only against the actual names of the backing columns. So,
for example, if someone had an alias for the column "id", so that it
would display as "identifier" when using this alias, we would validate
that the relationship defined with that field was defined in the config
using the name "id", and would throw a validation error if the config
instead used, "identifier."

## How was this tested?

Current test suite passing, along with a regression test for DwSql,
MsSql, MySql, and PostreSql.

## Sample Request(s)

Please see this issue for details on samples to reproduce.

https://github.com/Azure/data-api-builder/issues/2166

---------

Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>
Co-authored-by: Ayush Agarwal <34566234+ayush3797@users.noreply.github.com>