Handle Exception during dab validate when SqlMetadataProvider has not correctly initialized (#2202)
## Why make this change?
Closes https://github.com/Azure/data-api-builder/issues/2184
Closes https://github.com/Azure/data-api-builder/issues/2180


## What is this change?

When we don't properly initialize the `SqlMetadataProvider`, certain
functions will not behave as intended. This change adds some defensive
programming around this fact so that we have more readable exceptions in
these cases, as well as restructuring the validation to avoid the
potential of such a circumstance.

We refactor the `ValidateEntitiesMetadata` function into 2 functions,
one called `ValidateRelationshipConfigCorrectness` that validates the
relationships in the config without needing to cross reference database
metadata, and leaving the remaining validations that do require such
metadata within the original function. This validation is then handled
by a mix of the initialization of the metadata provider (this init
process does some validation), and the call to `ValidateRelationships`
which happens from within `ValidateEntitiesMetadata` only when the
initialization of the metadata provider resulted in no connection
errors.

Therefore, the normal code path when doing `dab validate` will be

1. Call `ValidateRelationshipConfigCorrectness`
2. Call `ValidateEntitiesMetadata`
    2a. If there were no connection errors Call `ValidateRelationships`



![image](https://github.com/Azure/data-api-builder/assets/93220300/1e2feec5-ac73-4e57-b3eb-01f032f362cf)


## How was this tested?

Manually, against our test suite, and we add a regression test that
looks for the correct error messaging when the config that caused the
error from the finding of the bug is used.

## Sample Request(s)

This can be reprod by using the command `dab validate -c <path to
config>