Fixing DateTime null parse issue in datawarehouse queries. (#2231)
## Why make this change?
When we moved to the new form of query execution:
https://github.com/Azure/data-api-builder/pull/2068, we got rid of the
resolver middleware that had the function:
`        public static bool RepresentsNullValue(JsonElement element)
        {
return (string.IsNullOrEmpty(element.ToString()) && element.GetRawText()
== "null");
        }`

which would check if the returned string was null. This was useful for
datetime values in datawarehouse where because of the use of string agg
we have to return null as a string.

The error that is seen is: "The string 'null' was not recognized as a
valid DateTime. There is an unknown word starting at index '0'.
## What is this change?
This change switches to leveraging tryparse for the datetime field
value. This allows us to handle the null case, as if the string is null,
tryparse will just fail and return null for date.

## How was this tested?
A test has been added to ensure that datetime null case is handled.
Before:
Running a sample query which checks for a date column:
`query
{
  tests
  {
    items
    {
      ReleaseDate
    }
  }
}`

![image](https://github.com/Azure/data-api-builder/assets/124841904/e29516b4-a61f-4805-bb9e-4fd942786185)

After fix:


![image](https://github.com/Azure/data-api-builder/assets/124841904/fb2cb85b-6188-49f9-92f4-56b6fb0ed7b0)

---------

Co-authored-by: Ayush Agarwal <34566234+ayush3797@users.noreply.github.com>