Replace internal references to RuntimeConfig with immutable copies to maintain consistent states (#2337)
## Why make this change?

closes https://github.com/Azure/data-api-builder/issues/2321

## What is this change?

Hot-Reload will mean refreshing the `RuntimeConfig` that the
`RuntimeConfigProvider` retrieves. To make sure that we have consistency
throughout DAB, we refactor how certain classes store the
`RuntimeConfig` to guarantee consistency within that class and with DAB.
For the `CosmosSqlMetadataProvider` this means that we make a copy of
the Entities at the time of construction and store that copy for use
within the class. We also save the state of the `isDevelopmentMode`
function from the runtimeConfig to keep this class consistent to the
`RuntimeConfig` at time of construction. This guarantees that no other
class can modify those Entities and we will have a consistent state
throughout the life of the `CosmosSqlMetadataProvider`. For the
`OpenApiDocumentor` we change the class to get a fresh `RuntimeConfig`
when it's public facing functions are called so that the output of those
functions matches the current state of DAB.

## How was this tested?

Against our current test suite. This will be tested from a hot-reload
perspective with the hot-reload integration tests.

## Sample Request(s)

Any request will work.