Refactor Config Loader and Provider for Hot-Reload (#2356)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/1966

## What is this change?

Move the `RuntimeConfig` to the `RuntimeConfigLoader`, which is now
responsible for maintaining that class in the proper state. We also move
the Hot-Reload function to this class, since it now has the
responsibility of maintaining a correct `RuntimeConfig`.

The provider therefore needs to have a `RuntimeConfigLoader` so that it
can access the `RuntimeConfig` through that class.

Initially, we were planning on splitting the `RuntimeConfigProvider`
into 2 classes that share a common interface: one for hosted scenarios,
and one for local. However, it became clear that this would create
additional problems with identifying the correct version of the provider
needed in the registered services. Please see the above related issue
for more information.

The `RuntimeConfigProvider` will maintain private access to the
`RuntimeConfigLoader`, however, the `RuntimeConfigLoader` will have
public access to its `RuntimeConfig`. This allows the current functions
to work with minimal changes, since we utilize out parameters in the
parsing and loading of the config, which then leads to certain functions
that modify the `RuntimeConfig` during initialization such as in the
Hosted scenario when we call the Provider's initialize function.

This will eventually be refactored so that the only class responsible
for modifying the `RuntimeConfig` is the loader, and we have an issue to
track this work here:
https://github.com/Azure/data-api-builder/issues/2362

## How was this tested?

Against current test suite.

## Sample Request(s)

Any request is valid.

---------

Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>