Add log-level property to config schema (#2359)
## Why make this change?

Adds log-level property to dab config file under the runtime section,
closing issue #1645. The log-level is necessary to receive updates from
the program, and with this property, it gives the users the ability to
change what type of information they receive from the program. Lastly,
it allows the users to have more options on how to change the log-level
property, be it through the config file or the CLI.

## What is this change?

First, the log-level property was added to the config file schema, in
order for the user to have the ability to add the property into their
config file. In order to save the information that is inside the
property, a new object model named `LogLevelOptions` was created, inside
of it is a new object model named `Level` with the purpose of showing
all of the possible values in an enum type. A `LogLevelOptions` was then
added to `RuntimeOptions` so the program knows in which section it is
supposed to parse the information. A new converter
`LogLevelOptionsConverterFactory` was created in order to allow the
property in the config file to be deserialized. The logic in `Startup`
file was modified to first tries to set the loggers based on the
log-level property, in the case that log-level is null, it will fall
back on the logic that was set up for `Host Mode`. New tests were
created in `ConfigurationTests` to test the validity of the code.

## How was this tested?

- [x] Integration Tests
- [ ] Unit Tests

---------

Co-authored-by: Ruben Cerna <rcernaserna@microsoft.com>