Add Support for Hot Reload of DataSource Section of Config file. (#2417)
## Why make this change?

- Closes #2323 

- **Improved Developer Experience:** Seamless updates allow developers
to update the database type or connection string without restarting the
application.
- **Enhanced Flexibility:** The application can dynamically switch
between different databases based on current needs or environment such
as database migration.
- **Increased Productivity:** Reduced downtime by eliminating the need
to restart the application for data-source changes.

## What is this change?
Supporting hot-reload for the data-source section involved two major
parts:

### Part 1: Refreshing Dependencies (PR:#2391)
The first part focused on supporting the refreshing of various
dependencies that rely on the DataSource section during startup.
However, these dependencies were not being refreshed upon changes to the
data-source configuration. This task involved adding the logic to
refresh those components that are configured on startup that acts as a
pre-requirement for Data-Source reload.

This critical part of the full support for hot-reloading was completed
in PR: #2391 by Aaron. This PR ensured that all necessary components are
refreshed when the data-source configuration changes.

### Part 2: Running Initialization with Updated Components/Dependencies
The second part involved running the initialization process once all
dependencies are updated. This initialization is crucial as it uses the
new data-source configuration to set up the connection and entity
refresh.

Key Changes:

Ensured that the initialization process is triggered after all
dependencies are refreshed and the same DataSourceName is maintained.
Added the initialization logic to refresh the new data-source
configuration.

## How was this tested?

- [x] Manual Tests since pipelines do not support switching between
databases.

### Testing Scenarios

#### Scenario A: Handling Invalid and Valid Passwords
Objective: To verify the system's behavior when using invalid and valid
passwords in the connection string.

Steps:
1. Start with MSSQL with a valid string.
2. update it to an incorrect credential in the connection string.
3. Show the logs to capture error messages.
4. Run the query and observe the failure.
5. Update the connection string with a valid credential.
6. Run the query again and observe the success.

Results:
Logs clearly indicated an invalid password error when using the
incorrect password.
The query failed with an invalid password for DB.
The query succeeded with a valid password for DB. 


#### Scenario B: Switching Between Databases
Objective: To ensure seamless transitions and query execution across
different databases.

Steps:

1. Start with MSSQL.
2. Run a query and observe the results.
3. Update the configuration to use MySQL.
4. Run the same query and observe the results.
5. Update the configuration to use PostgreSQL.
6. Run the same query and observe the results.

Observed Results:
The query executed successfully in MSSQL.
The query executed successfully in MySQL after updating the
configuration.
The query executed successfully in PostgreSQL after updating the
configuration.

---------

Co-authored-by: aaron burtle <aaronburtle@microsoft.com>