Added End-To-End Test for Datasource Hot-Reload (#2453)
## Why make this change?

This change solves the issue #1866

## What is this change?

Added an end-to-end test to ensure that DAB hot-reloads appropriately
when there is a connection change from one database to another database
(e.g. MSSQL to PostgreSQL).
In order to make tests related to hot reload less flaky, we stopped
using the sleep() function and added an event in
`FileSystemRuntimeConfigLoader`, which triggers when hot reload is
complete. This event is used by the tests in order to wait until hot
reload finishes.

## How was this tested?

- [X] Integration Tests
- [ ] Unit Tests

---------

Co-authored-by: Ruben Cerna <rcernaserna@microsoft.com>