Do not allow Mode to Hot-Reload (#2478)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/2457

## What is this change?

We add in an optional parameter to the `TryLoadConfig` function so that
when we Hot-Reload, we can pass along the `Mode`. When that parameter is
included (and therefore non null), we overwrite the Mode of the
`RuntimeConfig` to match the value of the previous `RuntimeConfig`. This
means that this Mode needs to have a public setter however, which we now
utilize to overwrite.

## How was this tested?

Modified our Unit test to provide coverage.

## Sample Request(s)

Any hot-reload that tries to change the mode will demonstrate the
functionality. So start the engine in dev mode, so that the filewatcher
will pick up any changes as we dont currently support hot-reload in
production, then change the Host Mode from development to production,
and note that the RuntimeConfig in-memory object is still in dev mode.

---------

Co-authored-by: Abhishek Kumar <abhishekkuma@microsoft.com>